[Fixes kbss-cvut/fta-fmea-ui#673] Set validator supporting User entity in UserRepositoryService - refactor -  AbstractEntityValidator from NamedEntityValidator - add UserValidator
