+++ src/main/java/cz/cvut/kbss/analysis/service/validation/DefaultEntityValidator.java
@@ -4,13 +4,14 @@ import cz.cvut.kbss.analysis.dao.BaseDao;
 import cz.cvut.kbss.analysis.model.NamedEntity;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.beanvalidation.SpringValidatorAdapter;
 
 @Qualifier("defaultEntityValidator")
 @Slf4j
-@org.springframework.stereotype.Component
-public class DefaultEntityValidator extends NamedEntityValidator{
+@Component
+public class DefaultEntityValidator extends NamedEntityValidator<NamedEntity>{
 
     public DefaultEntityValidator(SpringValidatorAdapter validatorAdapter) {
         super(NamedEntity.class, validatorAdapter);
@@ -22,7 +23,7 @@ public class DefaultEntityValidator extends NamedEntityValidator{
     }
 
     @Override
-    protected void customValidation(Object target, Errors errors, Object... validationHints) {
+    protected void customValidation(NamedEntity target, Errors errors, Object... validationHints) {
         // No custom validation
     }
 }
